escape function to escape strings together with trim and strip tags

+++ admin/includes/view_all_comments.php
@@ -35,39 +35,39 @@
       echo "<td>{$comment_author}</td>";
       echo "<td><div class='rj-td-wrap'>{$comment_content}</div></td>";
 
- 
+
 
 
 
       echo "<td>{$comment_email}</td>";
       echo "<td>{$comment_status}</td>";
 
-        $query = "SELECT * FROM posts WHERE post_id = $comment_post_id";
-        $select_post_id_query = mysqli_query($conn,$query);
-        while($row = mysqli_fetch_assoc($select_post_id_query)) {
-          $post_id = $row['post_id'];
-          $post_title = $row['post_title'];
+      $query = "SELECT * FROM posts WHERE post_id = $comment_post_id";
+      $select_post_id_query = mysqli_query($conn, $query);
+      while ($row = mysqli_fetch_assoc($select_post_id_query)) {
+        $post_id = $row['post_id'];
+        $post_title = $row['post_title'];
 
-          echo "<td><a href='../post.php?p_id=$post_id'>$post_title</a></td?";
-        }
+        echo "<td><a href='../post.php?p_id=$post_id'>$post_title</a></td?";
+      }
 
       echo "<td>{$comment_post_id}</td>";
-      
+
       echo "<td>{$comment_date}</td>";
 
-      
-    
+
+
       echo "<td >";
       echo "<a href='comments.php?approve={$comment_id}' class='';>Approve</a>";
 
       echo " / ";
       echo "<a href='comments.php?unapprove={$comment_id}' class='';>Unapprove</a></td>";
-      
-      
-    
+
+
+
       echo "<td style='display: flex;justify-content: center;align-items: center;background: red;'>";
       echo "<a href='comments.php?delete={$comment_id}' class='';>Delete</a></td>";
-      
+
       echo "</tr>";
     }
 
@@ -76,48 +76,53 @@
 
   </tbody>
 
-  <?php 
-  
-  
-  if(isset($_GET['approve'])) {
+  <?php
 
-    $approve_comment_id = $_GET['approve'];
-    
+
+  if (isset($_GET['approve'])) {
+
+    $approve_comment_id = escape($_GET['approve']);
+
     $query = "UPDATE comments SET comment_status = 'approved' WHERE comment_id = {$approve_comment_id}";
 
-    $status_query = mysqli_query($conn ,$query);
+    $status_query = mysqli_query($conn, $query);
 
     header("Location: comments.php");
-    
   }
-  
-  if(isset($_GET['unapprove'])) {
 
-    $unapprove_comment_id = $_GET['unapprove'];
-    
+  if (isset($_GET['unapprove'])) {
+
+    $unapprove_comment_id = escape($_GET['unapprove']);
+
     $query = "UPDATE comments SET comment_status = 'unapproved' WHERE comment_id = {$unapprove_comment_id}";
-    
-    $status_query = mysqli_query($conn ,$query);
+
+    $status_query = mysqli_query($conn, $query);
 
     header("Location: comments.php");
-
   }
-  
-  if(isset($_GET['delete'])) {
 
-    $del_comment_id = $_GET['delete'];
 
-    $query = "DELETE FROM comments WHERE comment_id = {$del_comment_id}";
+  if (isset($_GET['delete'])) {
 
-    $delete_query = mysqli_query($conn ,$query);
+    if (isset($_SESSION['user_role'])) {
 
-    header("Location: comments.php");
+      if ($_SESSION['user_role'] == 'admin') {
 
+
+        $del_comment_id = escape($_GET['delete']);
+
+        $query = "DELETE FROM comments WHERE comment_id = {$del_comment_id}";
+
+        $delete_query = mysqli_query($conn, $query);
+
+        header("Location: comments.php");
+      }
+    }
   }
-  
-  
-  
-  
+
+
+
+
   ?>
